+++ src/main/java/net/snowflake/client/core/SFLoginInput.java
@@ -4,9 +4,14 @@ import java.net.MalformedURLException;
 import java.net.URL;
 import java.security.PrivateKey;
 import java.time.Duration;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 import net.snowflake.client.core.auth.wif.WorkloadIdentityAttestation;
 import net.snowflake.client.jdbc.ErrorCode;
+import net.snowflake.client.jdbc.SnowflakeUtil;
 import org.apache.http.client.methods.HttpRequestBase;
 
 /** A class for holding all information required for login */
@@ -65,7 +70,7 @@ public class SFLoginInput {
   private String workloadIdentityProvider;
   private WorkloadIdentityAttestation workloadIdentityAttestation;
   private String workloadIdentityEntraResource;
-  private String workloadIdentityImpersonationPath;
+  private List<String> workloadIdentityImpersonationPath = Collections.emptyList();
 
   // OAuth
   private int redirectUriPort = -1;
@@ -630,13 +635,19 @@ public class SFLoginInput {
     return this;
   }
 
-  public String getWorkloadIdentityImpersonationPath() {
+  public List<String> getWorkloadIdentityImpersonationPath() {
     return workloadIdentityImpersonationPath;
   }
 
   public SFLoginInput setWorkloadIdentityImpersonationPath(
       String workloadIdentityImpersonationPath) {
-    this.workloadIdentityImpersonationPath = workloadIdentityImpersonationPath;
+    if (!SnowflakeUtil.isNullOrEmpty(workloadIdentityImpersonationPath)) {
+      this.workloadIdentityImpersonationPath =
+          Arrays.stream(workloadIdentityImpersonationPath.split(","))
+              .map(String::trim)
+              .filter(s -> !s.isEmpty())
+              .collect(Collectors.toList());
+    }
     return this;
   }